Faculty Project submission update
If the project does not submit successfully, an email will not be sent and the user will be sent to the error page. If the project is successful but a confirmation email was not sent, they will be informed and will be given their  project id on the error page.